# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Это API для отслеживания времени работы с использованием таймеров",
        "title": "TimerTracker API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8083",
    "basePath": "/v1",
    "paths": {
...
        "/v1/work/delete/{taskID}": {
            "delete": {
                "description": "Удаляет работу из системы по её ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "works"
                ],
                "summary": "Удаляет работу из базы данных",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Work ID",
                        "name": "workID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "$ref": "#/definitions/types.Result"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/types.Errors"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/types.Errors"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/types.Errors"
                        }
                    }
                }
            }
        },
        "/v1/work/edit": {
            "put": {
                "description": "Редактирует существующею работу в системе",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "works"
                ],
                "summary": "Редактирует работу в базе данных",
                "parameters": [
                    {
                        "description": "Объект работы, которую необходимо обновить",
                        "name": "task",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.Work"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.Work"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/types.Errors"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/types.Errors"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/types.Errors"
                        }
                    }
                }
            }
        },
        "/v1/work/period/{userID}": {
            "get": {
                "description": "Получает работы пользователя за указанный период времени",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "works"
                ],
                "summary": "Получает работы пользователя за период",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "userID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "format": "date-time",
                        "description": "Start Date Пример: 2023-10-05T21:48:00+07:00",
                        "name": "startDate",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "format": "date-time",
                        "description": "End Date Пример: 2023-10-05T21:48:00+07:00",
                        "name": "endDate",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/types.WorkPeriod"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/types.Errors"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/types.Errors"
                        }
                    },
                    "505": {
                        "description": "HTTP Version Not Supported",
                        "schema": {
                            "$ref": "#/definitions/types.Errors"
                        }
                    }
                }
            }
        },
        "/v1/work/{workID}": {
            "get": {
                "description": "Получает работу из системы по её ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "works"
                ],
                "summary": "Получает работу из базы данных по ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Work ID",
                        "name": "workID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.Work"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/types.Errors"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/types.Errors"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/types.Errors"
                        }
                    }
                }
            }
        },
        "/v1/works": {
            "get": {
                "description": "Получает список всех работ в системе",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "works"
                ],
                "summary": "Получает все работы из базы данных",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/types.Work"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/types.Errors"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "types.Errors": {
            "type": "object",
            "properties": {
                "errors": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "types.Result": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
...
        "types.Work": {
            "type": "object",
            "properties": {
                "date": {
                    "type": "string"
                },
                "hours": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "minutes": {
                    "type": "integer"
                },
                "taskID": {
                    "type": "integer"
                },
                "userID": {
                    "type": "integer"
                }
            }
        },
        "types.WorkPeriod": {
            "type": "object",
            "properties": {
                "endDate": {
                    "type": "string"
                },
                "startDate": {
                    "type": "string"
                },
                "totalHours": {
                    "type": "integer"
                },
                "totalMinutes": {
                    "type": "integer"
                },
                "totalMinutesCombined": {
                    "type": "integer"
                },
                "userID": {
                    "type": "integer"
                }
            }
        }
    }
}